Working 3 File System
I moved the login type stuff to a seperate file to furthen illustrate how to use modules.  The main issue with your original code was that you had main methods and other stuff running in your modules which generally you do not want.  Keep in mind that modules are interpreted when they are imported.